requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Pillow==11.0.0
